Initial commit of CADR HDL.
This isn't complete yet and will not work.  In particular, the block storage
device is unimplemented and the RAM controller probably violates the XBUS
requirements.  Additionally, I would like to take advantage of the large amount
of block ram on the MiSTer by making a cache (probably ~64 KW in size).