Fix DDRAM controller, make testbench, and add promdisable (still no boot)
The DDRAM controller had a bug that caused reads to hang (it improperly waited
on DDRAM_BUSY when starting a read, but DDRAM_BUSY is only set for writes).

This fix wasn't sufficient to allow proper operation -- one potential problem
is that the VRAM controller fails in simulation (I'm investigating this), but
this is the same logic used by the LX45, which does work there (however,
differences in the block RAM IPs could cause problems, so this is under
investigation).

I also plan to figure out how to get the spy port to work, which should help
debugging real hardware.